+++ capa-controller-role/giantswarm-capa-role.tf
@@ -32,6 +32,18 @@ resource "aws_iam_role_policy_attachment" "giantswarm-capa-controller-policy-att
   policy_arn = aws_iam_policy.giantswarm-capa-controller-policy.arn
 }
 
+resource "aws_iam_policy" "giantswarm-capa-controller-policy-vpc" {
+  count  = var.byovpc ? 0 : 1 # This policy is not needed in BYO VPC installations
+  name   = "giantswarm-${var.installation_name}-capa-controller-policy-vpc"
+  policy = file("${path.module}/capa-controller-policy-vpc.json")
+  tags   = local.tags
+}
+resource "aws_iam_role_policy_attachment" "giantswarm-capa-controller-policy-vpc-attachment" {
+  count      = var.byovpc ? 0 : 1 # This policy is not needed in BYO VPC installations
+  role       = aws_iam_role.giantswarm-capa-controller-role.name
+  policy_arn = aws_iam_policy.giantswarm-capa-controller-policy-vpc[0].arn
+}
+
 resource "aws_iam_policy" "giantswarm-dns-controller-policy" {
   name   = "giantswarm-${var.installation_name}-dns-controller-policy"
   policy = file("${path.module}/dns-controller-policy.json")