+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Add support for removing some IAM permissions from the capa controller role in BYOVPC installations.
+
 ## [4.2.0] - 2024-09-04
 
 ### Changed